Updated hat menu, ghosts are back, bugfixes
Updated the hat menu to allow for scaling and easily adding more hats. Added a cactus-hat while I'm figuring out how Blender works, will update with more hats soon. Also reenabled ghosts, needs testing. Bugfixes with parsing floats from backend, there is some unintended behavior with decimalpoint separation. Also added some animations to the menu.

+++ Assets/HatManager.cs
@@ -26,10 +26,13 @@ public class HatManager : MonoBehaviour
         if (hatExists)
         {
             RemoveHat();
-            PlayerPrefs.SetString("hat", "no");
-            PlayerPrefs.DeleteKey("hatindex");
-            PlayerPrefs.Save();
-
+            if (PlayerPrefs.GetInt("hatindex") == hatIndex) return;
+            else
+            {
+                PlayerPrefs.SetString("hat", "no");
+                PlayerPrefs.DeleteKey("hatindex");
+                PlayerPrefs.Save();
+            }
         }
 
         if (hatIndex >= 0 && hatIndex < hatPrefabs.Count)
@@ -51,7 +54,7 @@ public class HatManager : MonoBehaviour
 
             currentHats.Add(hat);
             string hatName = hatPrefab.name.Replace("(Clone)", "");
-            Debug.Log("Set " + hatName + " to playerprefs");
+
             PlayerPrefs.SetString("hat", hatName);
             PlayerPrefs.Save();
             PlayerPrefs.SetInt("hatindex", hatIndex);
